+++ CHANGELOG.md
@@ -4,11 +4,13 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Added
 - [#1414] Cancelling in-progress exportation
+- [#1755] Copy tags to tag clipboard and paste them elsewhere
 
 ### Fixed
 - FLA export printing xxx string on exporting character with id 320
 - Copy to with dependencies does not refresh timeline
 - Copy to with dependencies does not set the timelined, that can result to missing dependencies (red tags in the tree)
+- Double warning/error when copy to / move to and same character id already exists
 
 ## [16.1.0] - 2022-11-06
 ### Added
@@ -2511,6 +2513,7 @@ All notable changes to this project will be documented in this file.
 [alpha 8]: https://github.com/jindrapetrik/jpexs-decompiler/compare/alpha7...alpha8
 [alpha 7]: https://github.com/jindrapetrik/jpexs-decompiler/releases/tag/alpha7
 [#1414]: https://www.free-decompiler.com/flash/issues/1414
+[#1755]: https://www.free-decompiler.com/flash/issues/1755
 [#1459]: https://www.free-decompiler.com/flash/issues/1459
 [#1832]: https://www.free-decompiler.com/flash/issues/1832
 [#1849]: https://www.free-decompiler.com/flash/issues/1849